probes/requirements.txt
requests==2.32.3
confluent-kafka>=2.14.0
python-dotenv==1.0.1
